+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/DirectorDaoImpl.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.dao.DirectorDao;
 import ru.yandex.practicum.filmorate.model.Director;
-import ru.yandex.practicum.filmorate.util.DirectorUtil;
+import ru.yandex.practicum.filmorate.util.Director.DirectorUtil;
 
 import java.util.List;
 import java.util.Optional;
