+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/FriendDaoImpl.java
@@ -5,7 +5,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.util.UserUtil;
+import ru.yandex.practicum.filmorate.util.user.UserUtil;
 import ru.yandex.practicum.filmorate.dao.FriendDao;
 import ru.yandex.practicum.filmorate.model.User;
